Moving keen dependent stuff to dataset.keen file
