+++ hyperbolic-rs/rudders/datasets/keen.py
@@ -14,7 +14,9 @@
 """File with keen dataset specific functions"""
 
 import json
-KEEN_METADATA = "data/keen/keens_and_gems_25_june_2020.json"
+import re
+URL_RE = '((www\.[^\s]+)|(https?://[^\s]+)|(http?://[^\s]+))'
+KEEN_METADATA = "data/keen/exports_2020-07-03_keens_and_gems.jsonl"
 USER_ITEM_INTERACTIONS_FILE = "interactions.csv"
 
 
@@ -64,3 +66,56 @@ def build_iid2title(item_id_key, item_title_key):
             if item_id_key in line and item_title_key in line:
                 iid2title[line[item_id_key]] = line[item_title_key][1:-1]
     return iid2title
+
+
+def get_keens(data):
+    all_keens = {}
+    for item in data:
+        keen_id = item["keen_id"]
+        gem = None
+        if "gem_id" in item:
+            gem = Gem(item)
+
+        if keen_id in all_keens:
+            keen = all_keens[keen_id]
+        else:
+            keen = Keen(item["keen_id"], item["keen_title"], get_value(item, "keen_description"),
+                        item["keen_creator_uid"])
+            all_keens[keen_id] = keen
+
+        if gem is not None and not gem.is_empty():
+            keen.gems.append(gem)
+    return all_keens
+
+
+def process_input(string):
+    string = string[1:-1] if type(string) == str else ""
+    return string.replace("\\n", "")
+
+
+def get_value(item, key):
+    return item[key] if key in item else ""
+
+
+class Keen:
+    def __init__(self, keen_id, title, description, creator_uid):
+        self.keen_id = keen_id
+        self.title = process_input(title)
+        self.description = process_input(description)
+        self.creator_uid = process_input(creator_uid)
+        self.gems = []
+
+
+class Gem:
+    def __init__(self, item):
+        self.gem_id = get_value(item, "gem_id")
+        text = get_value(item, "gem_text")
+        text = re.sub(URL_RE, '', text)  # replace URL for empty string
+        self.text = process_input(text)
+        self.link_url = process_input(get_value(item, "gem_link_url"))
+        self.link_title = process_input(get_value(item, "gem_link_title"))
+        self.link_description = process_input(get_value(item, "gem_link_description"))
+        self.creator_uid = process_input(get_value(item, "gem_uid"))
+
+    def is_empty(self):
+        return not self.text and not self.link_url and not self.link_title and not self.link_description
